# packages/shared-types/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/businesses/me": {
      "get": {
        "operationId": "BusinessesController_getCurrentBusiness",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current business profile for the tenant context.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "No business exists for the current tenant context."
          }
        },
        "summary": "Get the current tenant business profile",
        "tags": [
          "Businesses"
        ]
      }
    },
...
  },
  "info": {
    "title": "Quickarte API",
    "description": "",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001/v1"
    }
  ],
  "components": {
    "schemas": {
      "BusinessResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "currency": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "locale": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "type",
          "currency",
          "timezone",
          "locale"
        ]
      },
...
    }
  }
}